requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
django-prometheus==2.3.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
prometheus_client==0.21.1
PyJWT==1.7.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.2
urllib3==2.2.3
uvicorn==0.34.0
